+++ test/bdd-syntax/steps-cucumber-style/fixtures.ts
@@ -2,7 +2,7 @@ import { test as base, createBdd } from 'playwright-bdd';
 import { World } from './world';
 
 export const test = base.extend<{ world: World }>({
-  world: ({ $tags }, use, testInfo) => use({ testInfo, tags: $tags }),
+  world: ({ $tags, $step }, use, testInfo) => use({ testInfo, tags: $tags, step: $step }),
 });
 
 export const { Given, When, Then } = createBdd(test, { worldFixture: 'world' });