+++ test/bdd-syntax/steps-cucumber-style/steps.ts
@@ -38,9 +38,13 @@ Then('Custom param-no-transformer is {param-no-transformer}', function (param: s
   expect(param).toEqual('dog');
 });
 
-Then('Passed doc string to contain {string}', async function (arg: string, docString: string) {
-  expect(docString).toContain(arg);
-});
+Then(
+  'Passed doc string has media type {string} and contains {string}',
+  function (mediaType: string, content: string, docString: string) {
+    expect(this.step.docStringType).toEqual(mediaType);
+    expect(docString).toContain(content);
+  },
+);
 
 Then(
   'Passed data table to have in row {int} col {string} value {string}',